Better error messages when raising exceptions.
So far, when we HTTP status is not successful, we already raise a
contextualized exception. However, only that is not enough to easily
debug the issue, since the actual HTTP status and body message are not
known.

This aims to make it easier to debug unsuccessful API responses by
providing more information when raising exceptions.